metrics_producer/requirements.txt
kafka-python==1.4.7
psutil==5.6.7
twisted==19.10.0
